backend/requirements.txt
Django==4.2.9
psycopg2-binary>=2.9
djangorestframework>=3.14.0
Pillow>=10.2.0
django-cors-headers==3.8.0
requests==2.31.0
channels[daphne]>=3.1.2
numpy==1.26.4
websockets==12.0
django-sslserver==0.22
Twisted[tls,http2]>=22.3.0
adrf==0.1.5